pyproject.toml
[tool.poetry]
name = "illust-exporter"
version = "0.1.0"
description = "A tool to export illustrations for specific sites."
authors = ["kittyflute1812 <kittyflute1812@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9, <3.13"
Pillow = "*"
psd-tools = "*"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[tool.poetry.scripts]
illust-exporter = "src.illust_exporter:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
